analytics-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1